Improve style of array type

+++ Sources/LLVMSwift/LLVMArrayType.swift
@@ -2,10 +2,9 @@ import LLVM
 
 public class LLVMArrayType<Element: LLVMValue>: LLVMType<LLVMArray<Element>> {
     
-    public typealias ElementCount = Int
+    public typealias ElementCount = UInt64
     
     @inlinable public convenience init(elementType: LLVMType<Element>, elementCount: ElementCount) {
-        self.init(rawType: LLVMArrayType2(elementType.rawType, UInt64(elementCount)))
+        self.init(rawType: LLVMArrayType2(elementType.rawType, elementCount))
     }
 }
-
